partition_manager: Add generated alignment partitions to parent spans
When an empty partition is created to fulfill an alignment requirement,
it is not added to the span of the partition being aligned.

This commit adds the empty partitions to any spans
that contain the partition being aligned, as long as it doesn't
change the size or start/end addresses of the span.

Ref: NCSDK-3997

Signed-off-by: Jonathan Nilsen <Jonathan.Nilsen@nordicsemi.no>