tests: benchmarks: idle_gpio: set PPR input pin to NS
Configure the input pin used by PPR in the idle_gpio testcase as
nonsecure, which is required for it to be able to trigger the
nonsecure GPIOTE interrupt that is connected to PPR.

This should be considered a minimal workaround for enabling the
test, and will be superseded by a more proper setup.
The workaround enables the PERIPHCONF section in the PPR build and
uses PERIPHCONF macros directly to set the secure attribute of the
input pin. This works as long as no other build tries to configure
the pin, which could lead to a conflict in build time.

Signed-off-by: Jonathan Nilsen <jonathan.nilsen@nordicsemi.no>